requirements.txt
seaborn
pandas
scikit-learn
gcsfs
google-cloud-storage
fsspec
